requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
numpy==1.26.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0


